Create TaxCalculator
This program is a simple tax calculator that calculates the weekly tax withholding based on a user's income. The program first prompts the user to enter their weekly income, and then uses a series of if statements to determine the appropriate tax rate. The tax rate is based on the following thresholds:

Income less than $500: 10%
Income between $500 and $1,500: 15%
Income between $1,500 and $2,500: 20%
Income greater than $2,500: 30%